+++ CHANGELOG.md
@@ -17,5 +17,6 @@ All notable changes to this project will be documented in this file.
 
 - Changed slurm-client version to 0.3.1
 - Changed references from V0041 to V0043 data parser in pod controller
+- Changed kind.sh to use Slurm dynamic nodes
 
 ### Removed